+++ frontend/src/app/expected-result-comparison/ExpectedResultComparison.common.scss
