+++ frontend/src/app/expected-result-comparison/ExpectedResultComparison.dark.scss
