+++ frontend/src/app/expected-result-comparison/ExpectedResultComparison.light.scss
