+++ frontend/src/app/expected-result-comparison/models/ComparisonResult.ts
