+++ frontend/src/app/expected-result-comparison/models/ExpectedVsTimeSeriesRequest.ts
