+++ frontend/src/app/expected-result-comparison/models/SimulationVsExpectedRequest.ts
