+++ frontend/src/app/expected-result-comparison/models/SimulationVsTimeSeriesRequest.ts
