+++ frontend/src/app/expected-result-comparison/models/TimeSeriesVsTimeSeriesRequest.ts
