+++ frontend/src/app/expected-result-comparison/views/expected-vs-time-series/ExpectedVsTimeSeries.light.scss
