+++ frontend/src/app/expected-result-comparison/views/result-viewer/ResultViewer.common.scss
