+++ frontend/src/app/expected-result-comparison/views/simulation-vs-expected/SimulationVsExpected.dark.scss
