+++ frontend/src/app/expected-result-comparison/views/simulation-vs-time-series/SimulationVsTimeSeries.dark.scss
