+++ frontend/src/app/expected-result-comparison/views/time-series-vs-time-series/TimeSeriesVsTimeSeriesChartResult.dark.scss
